requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.15
psycopg2-binary>=2.9.9
tenacity>=8.2.3 